Second review iteration after fixes implementation

+++ contracts/TestStorageExample/StorageVersionUpgradeExample.sol
@@ -1,15 +1,17 @@
 pragma solidity ^0.5.0;
 
-import './UpgradeableStorage.sol';
+import "./UpgradeableStorage.sol";
+
 
 contract Storage_V0 is UpgradeableStorage {
-  function test_storage_v0 () public view returns(uint) {
-    return instrumentCounter;
-  }
+    function test_storage_v0() public view returns (uint256) {
+        return instrumentCounter;
+    }
 }
 
+
 contract Storage_V1 is UpgradeableStorage {
- function test_storage_v1 () public view returns(uint) {
-    return instrumentCounter + 1;
-  }
+    function test_storage_v1() public view returns (uint256) {
+        return instrumentCounter + 1;
+    }
 }
